+++ test/all.dart
@@ -1,3 +1,4 @@
 library bwu_grinder_tasks.test.all;
 
-main() {}
+///
+void main() {}