+++ test/simple_test.dart
@@ -1,0 +1,19 @@
+import 'package:test/test.dart';
+
+///
+void main() {
+  group('', () {
+    setUp(() {});
+
+    tearDown(() {});
+
+    test('basic', () {
+      // set up
+//      var done = expectAsync(() {});
+      // exercise
+      // verification
+      expect('', equals(''));
+      // tear down
+    });
+  });
+}