+++ tool/grind.dart
@@ -4,7 +4,8 @@ export 'package:bwu_grinder_tasks/bwu_grinder_tasks.dart' hide main;
 import 'package:bwu_grinder_tasks/bwu_grinder_tasks.dart'
     show grind, analyzerIgnoreInfoMessages;
 
-main(List<String> args) {
+///
+void main(List<String> args) {
   // TODO(zoech) remove when deprecated `analyzer_task` and `dartformat_task` are removed
   analyzerIgnoreInfoMessages = true;
   grind(args);