+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # Changelog
 
 ## 0.1.1
-
+- the `analyze` task now uses Grinders `Analyzer.analyze()` instead of running `tuneup`.
+- the `test` task doesn't use `vm` and `content-shell` as default anymore. Use `dart_test.yaml` instead to configure
+  which tests to run. See https://github.com/dart-lang/test/blob/master/doc/configuration.md for details.
 - update to Grinder 0.8.0
-- fix strong-mode warnings 
+- fix strong-mode warnings
 
 ## 0.1.0
 
 - move Grinder tasks from bwu_utils_dev
-- remove direct dependency on bwu_archive_downloader and use `pub global run` 
+- remove direct dependency on bwu_archive_downloader and use `pub global run`
 instead.